Workaround for iOS 7 cross-fade bug on login screen POTD
Using a CABasicAnimation on the UIImageView 'contents' worked fine
on iOS 6, but on iOS 7 it was intermittently failing, turning into
a kind of binary flip or flash to white without a proper fade.

This patch switches from a single image view to two image views;
the upper one is shown only during the fades and its opacity is
animated instead of its contents. When the fade completes, the
image is transferred to the main image view and the fade view is
hidden.

Feels nasty, but seems to work.